fix: JSON fails checksum
Maps cannot be re-marshaled. It can cause differences between the source and target JSON. To avoid this
from happening maps are no longer a supported changeset column value. Any JSON must be passed as the
original string value. This bug was only affecting the notify listener. The logical replication listener
was already passing JSON as a string.