refactor: Change schema is target schema by default
This change is necessary to avoid potential data corruption. Unless overridden, the changes
schema should always be used when writing to the target. Since data integrity checks only
check to see if the two changesets tables on both the source and target are the same, it will
not detect if data is written to the wrong schema.